refactor(test): update call sites to import suppressGlobalAnnouncements directly
Remove the re-export from common.ts and update the four siem_migrations
.cy.ts files to import from suppress_global_announcements.ts directly.
