+++ x-pack/solutions/security/test/security_solution_cypress/cypress/e2e/investigations/siem_migrations/dashboards/onboarding.cy.ts
@@ -11,7 +11,8 @@ import {
   ONBOARDING_SIEM_MIGRATIONS_LIST,
   ONBOARDING_TRANSLATIONS_RESULT_TABLE,
 } from '../../../../screens/siem_migrations';
-import { deleteConnectors, suppressGlobalAnnouncements } from '../../../../tasks/api_calls/common';
+import { deleteConnectors } from '../../../../tasks/api_calls/common';
+import { suppressGlobalAnnouncements } from '../../../../tasks/api_calls/suppress_global_announcements';
 import { createBedrockConnector } from '../../../../tasks/api_calls/connectors';
 import { cleanDashboardsMigrationData } from '../../../../tasks/api_calls/siem_migrations';
 import { visit } from '../../../../tasks/navigation';
