+++ x-pack/solutions/security/test/security_solution_cypress/cypress/e2e/investigations/siem_migrations/rules/translated_rules_page.cy.ts
@@ -13,7 +13,8 @@ import {
   TRANSLATED_RULE_RESULT_BADGE,
   TRANSLATED_RULES_RESULT_TABLE,
 } from '../../../../screens/siem_migrations';
-import { deleteConnectors, suppressGlobalAnnouncements } from '../../../../tasks/api_calls/common';
+import { deleteConnectors } from '../../../../tasks/api_calls/common';
+import { suppressGlobalAnnouncements } from '../../../../tasks/api_calls/suppress_global_announcements';
 import { createBedrockConnector } from '../../../../tasks/api_calls/connectors';
 import { visit } from '../../../../tasks/navigation';
 import {
