+++ x-pack/solutions/security/test/security_solution_cypress/cypress/tasks/api_calls/common.ts
@@ -56,8 +56,6 @@ export const rootRequest = <T = unknown>({
   }
 };
 
-export { suppressGlobalAnnouncements } from './suppress_global_announcements';
-
 // a helper function to wait for the root request to be successful
 // defaults to 5 second intervals for 3 attempts
 // can be helpful when waiting for a resource to be created before proceeding
